Development

+++ server/db/models/wig.js
@@ -16,8 +16,11 @@ const Wig = db.define('wig', {
     allowNull: false
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false
+    // get() {
+    //   return this.getDataValue('price') / 100;
+    // }
   },
   quantity: {
     type: Sequelize.INTEGER,
@@ -67,7 +70,7 @@ Wig.findByIds = async function(wigIds) {
   return wigstoupdate;
 };
 
-// Instance method to check if a wig has sufficient quantity/stock to fulfill an order.
+// Instance method that returns true if there's sufficient stock/quantity to fulfill the order and false if there isn't.
 Wig.prototype.checkQuantity = function(orderQuantity) {
   if (this.quantity >= orderQuantity) {
     return true;
